backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openpyxl==3.1.2
alembic==1.13.1
bcrypt==4.1.2
